init.qcom.usb.sh: Check serialnumber value before writing dummy value
Currently .rc updates serialnumber from ro.serialno on boot and before
enabling .sh to execute. Once .sh starts executing, it tries to read
ro.serialno and if it shows empty string,  it tries to write dummy value
for adb interface to work properly. But it happens that .sh couldn't
access ro.serialno due to sepolicy permission and in that .sh endup
overriding serialnumber with dummy value.  This causes all devices
get same dummy value instead of having different serialnumber for each
hardware.  Fix the issue by checking serialnumber in .sh if it is already
updated by .rc rule or not and write dummy value to serialnumber sysfs entry
only if it is empty.

Change-Id: Ie07dfa6f05d9847800143b4c318c55367e865935